backend/requirements.txt
fastapi
uvicorn
azure-cosmos
python-dotenv
passlib[bcrypt]
pydantic